Update ADC driver according to new MATH functions. Use int32_t type for ADC data.

+++ inc/adc.h
@@ -102,33 +102,33 @@ ADC_status_t ADC_init(const ADC_gpio_t* pins);
 ADC_status_t ADC_de_init(void);
 
 /*!******************************************************************
- * \fn ADC_status_t ADC_convert_channel(ADC_channel_t channel, uint16_t* adc_data_12bits)
+ * \fn ADC_status_t ADC_convert_channel(ADC_channel_t channel, int32_t_t* adc_data_12bits)
  * \brief Perform a channel conversion.
  * \param[in]  	channel: Channel to convert.
- * \param[out] 	adc_data_12bits: Pointer to short that will contain the 12-bits ADC data.
+ * \param[out] 	adc_data_12bits: Pointer to integer that will contain the 12-bits ADC data.
  * \retval		Function execution status.
  *******************************************************************/
-ADC_status_t ADC_convert_channel(ADC_channel_t channel, uint16_t* adc_data_12bits);
+ADC_status_t ADC_convert_channel(ADC_channel_t channel, int32_t* adc_data_12bits);
 
 /*!******************************************************************
- * \fn ADC_status_t ADC_compute_vmcu(uint16_t ref_voltage_12bits, uint16_t ref_voltage_mv, int32_t* vmcu_mv)
+ * \fn ADC_status_t ADC_compute_vmcu(int32_t_t ref_voltage_12bits, int32_t_t ref_voltage_mv, int32_t* vmcu_mv)
  * \brief Compute MCU voltage.
  * \param[in]  	ref_voltage_12bits: Reference voltage 12-bits raw data from ADC.
  * \param[in]	ref_voltage_mv: Reference voltage in mV.
  * \param[out] 	vmcu_mv: Pointer to integer that will contain the MCU voltage in mV.
  * \retval		Function execution status.
  *******************************************************************/
-ADC_status_t ADC_compute_vmcu(uint16_t ref_voltage_12bits, uint16_t ref_voltage_mv, int32_t* vmcu_mv);
+ADC_status_t ADC_compute_vmcu(int32_t ref_voltage_12bits, int32_t ref_voltage_mv, int32_t* vmcu_mv);
 
 /*!******************************************************************
- * \fn ADC_status_t ADC_compute_tmcu(int32_t vmcu_mv, uint16_t tmcu_12bits, int32_t* tmcu_degrees)
+ * \fn ADC_status_t ADC_compute_tmcu(int32_t vmcu_mv, int32_t_t tmcu_12bits, int32_t* tmcu_degrees)
  * \brief Compute MCU temperature.
  * \param[in]	vmcu_mv: MCU supply voltage in mV.
  * \param[in]  	tmcu_12bits: Temperature sensor 12-bits raw data from ADC.
  * \param[out] 	tmcu_degrees: Pointer to integer that will contain MCU temperature in 2's complement format.
  * \retval		Function execution status.
  *******************************************************************/
-ADC_status_t ADC_compute_tmcu(int32_t vmcu_mv, uint16_t tmcu_12bits, int32_t* tmcu_degrees);
+ADC_status_t ADC_compute_tmcu(int32_t vmcu_mv, int32_t tmcu_12bits, int32_t* tmcu_degrees);
 
 /*!******************************************************************
  * \fn int32_t ADC_get_vrefint_voltage_mv(void)
